refactor: split the CoS prompt pre-step resolvers into cosTaskPreStepBlocks.js (#5695)
cosTaskGenerator.js carried two layers that churn independently: the
task-SELECTION engine (evaluateTasks, the spawnPriorityN* ladder, the
improvement/idle-review generators) and a prompt PRE-STEP layer — the
per-task-type resolvers that run a real pre-flight scan (branch reconcile, repo
sync, issue reconcile, reference-repo diff, gh PR poll), decide whether to
dispatch at all, and render the block the agent prompt is built around. Most of
the file's churn landed in the second, so every prompt tweak rewrote the same
file as every scheduling change.

The pre-step layer, the static prompt-fragment builders it substitutes (author
filter, exclude labels, swarm, plan constraint), the two perpetual-drain brakes
and the token renderer now live in server/services/cosTaskPreStepBlocks.js. The
move is verbatim — generated prompt text is byte-identical — and the deferred
`await import(...)` calls stay as they are (cycle avoidance, not laziness).
cosTaskGenerator.js keeps a back-compat re-export of the five helpers that were
public there, so deep imports on other installs and forks keep resolving.

One real fix rides along: `resolveRepoSyncBlock` called `isTruthyMeta` without
binding it (the name was destructured in a sibling resolver's dynamic import),
so a repo-sync task with `requireApproval` set threw a ReferenceError instead of
withholding the sweep. Nothing covered that path; it is bound now.
